+++ lib/Model/ResetUserNotifications200Response.php
@@ -62,8 +62,8 @@ class ResetUserNotifications200Response implements ModelInterface, ArrayAccess, 
         'code' => 'string',
         'reason' => 'string',
         'secondary_code' => 'string',
-        'banned_until' => 'float',
-        'max_character_length' => 'float',
+        'banned_until' => 'int',
+        'max_character_length' => 'int',
         'translated_error' => 'string',
         'custom_config' => '\FastComments\Client\Model\CustomConfigParameters'
     ];
@@ -80,8 +80,8 @@ class ResetUserNotifications200Response implements ModelInterface, ArrayAccess, 
         'code' => null,
         'reason' => null,
         'secondary_code' => null,
-        'banned_until' => 'double',
-        'max_character_length' => 'double',
+        'banned_until' => 'int64',
+        'max_character_length' => 'int32',
         'translated_error' => null,
         'custom_config' => null
     ];
@@ -459,7 +459,7 @@ class ResetUserNotifications200Response implements ModelInterface, ArrayAccess, 
     /**
      * Gets banned_until
      *
-     * @return float|null
+     * @return int|null
      */
     public function getBannedUntil()
     {
@@ -469,7 +469,7 @@ class ResetUserNotifications200Response implements ModelInterface, ArrayAccess, 
     /**
      * Sets banned_until
      *
-     * @param float|null $banned_until banned_until
+     * @param int|null $banned_until banned_until
      *
      * @return self
      */
@@ -486,7 +486,7 @@ class ResetUserNotifications200Response implements ModelInterface, ArrayAccess, 
     /**
      * Gets max_character_length
      *
-     * @return float|null
+     * @return int|null
      */
     public function getMaxCharacterLength()
     {
@@ -496,7 +496,7 @@ class ResetUserNotifications200Response implements ModelInterface, ArrayAccess, 
     /**
      * Sets max_character_length
      *
-     * @param float|null $max_character_length max_character_length
+     * @param int|null $max_character_length max_character_length
      *
      * @return self
      */
