+++ lib/Model/SaveComment200Response.php
@@ -65,8 +65,8 @@ class SaveComment200Response implements ModelInterface, ArrayAccess, \JsonSerial
         'reason' => 'string',
         'code' => 'string',
         'secondary_code' => 'string',
-        'banned_until' => 'float',
-        'max_character_length' => 'float',
+        'banned_until' => 'int',
+        'max_character_length' => 'int',
         'translated_error' => 'string',
         'custom_config' => '\FastComments\Client\Model\CustomConfigParameters'
     ];
@@ -86,8 +86,8 @@ class SaveComment200Response implements ModelInterface, ArrayAccess, \JsonSerial
         'reason' => null,
         'code' => null,
         'secondary_code' => null,
-        'banned_until' => 'double',
-        'max_character_length' => 'double',
+        'banned_until' => 'int64',
+        'max_character_length' => 'int32',
         'translated_error' => null,
         'custom_config' => null
     ];
@@ -574,7 +574,7 @@ class SaveComment200Response implements ModelInterface, ArrayAccess, \JsonSerial
     /**
      * Gets banned_until
      *
-     * @return float|null
+     * @return int|null
      */
     public function getBannedUntil()
     {
@@ -584,7 +584,7 @@ class SaveComment200Response implements ModelInterface, ArrayAccess, \JsonSerial
     /**
      * Sets banned_until
      *
-     * @param float|null $banned_until banned_until
+     * @param int|null $banned_until banned_until
      *
      * @return self
      */
@@ -601,7 +601,7 @@ class SaveComment200Response implements ModelInterface, ArrayAccess, \JsonSerial
     /**
      * Gets max_character_length
      *
-     * @return float|null
+     * @return int|null
      */
     public function getMaxCharacterLength()
     {
@@ -611,7 +611,7 @@ class SaveComment200Response implements ModelInterface, ArrayAccess, \JsonSerial
     /**
      * Sets max_character_length
      *
-     * @param float|null $max_character_length max_character_length
+     * @param int|null $max_character_length max_character_length
      *
      * @return self
      */
