+++ lib/Model/UpdateUserNotificationStatus200Response.php
@@ -59,14 +59,14 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
       */
     protected static $openAPITypes = [
         'status' => '\FastComments\Client\Model\ImportedAPIStatusFAILED',
-        'matched_count' => 'float',
-        'modified_count' => 'float',
+        'matched_count' => 'int',
+        'modified_count' => 'int',
         'note' => 'string',
         'reason' => 'string',
         'code' => 'string',
         'secondary_code' => 'string',
-        'banned_until' => 'float',
-        'max_character_length' => 'float',
+        'banned_until' => 'int',
+        'max_character_length' => 'int',
         'translated_error' => 'string',
         'custom_config' => '\FastComments\Client\Model\CustomConfigParameters'
     ];
@@ -80,14 +80,14 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
       */
     protected static $openAPIFormats = [
         'status' => null,
-        'matched_count' => 'double',
-        'modified_count' => 'double',
+        'matched_count' => 'int64',
+        'modified_count' => 'int64',
         'note' => null,
         'reason' => null,
         'code' => null,
         'secondary_code' => null,
-        'banned_until' => 'double',
-        'max_character_length' => 'double',
+        'banned_until' => 'int64',
+        'max_character_length' => 'int32',
         'translated_error' => null,
         'custom_config' => null
     ];
@@ -432,7 +432,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Gets matched_count
      *
-     * @return float
+     * @return int
      */
     public function getMatchedCount()
     {
@@ -442,7 +442,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Sets matched_count
      *
-     * @param float $matched_count matched_count
+     * @param int $matched_count matched_count
      *
      * @return self
      */
@@ -459,7 +459,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Gets modified_count
      *
-     * @return float
+     * @return int
      */
     public function getModifiedCount()
     {
@@ -469,7 +469,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Sets modified_count
      *
-     * @param float $modified_count modified_count
+     * @param int $modified_count modified_count
      *
      * @return self
      */
@@ -604,7 +604,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Gets banned_until
      *
-     * @return float|null
+     * @return int|null
      */
     public function getBannedUntil()
     {
@@ -614,7 +614,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Sets banned_until
      *
-     * @param float|null $banned_until banned_until
+     * @param int|null $banned_until banned_until
      *
      * @return self
      */
@@ -631,7 +631,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Gets max_character_length
      *
-     * @return float|null
+     * @return int|null
      */
     public function getMaxCharacterLength()
     {
@@ -641,7 +641,7 @@ class UpdateUserNotificationStatus200Response implements ModelInterface, ArrayAc
     /**
      * Sets max_character_length
      *
-     * @param float|null $max_character_length max_character_length
+     * @param int|null $max_character_length max_character_length
      *
      * @return self
      */
